+++ cli/src/commands/bookmark/untrack.rs
@@ -15,7 +15,7 @@
 use clap_complete::ArgValueCandidates;
 use itertools::Itertools as _;
 
-use super::find_remote_bookmarks;
+use super::find_trackable_remote_bookmarks;
 use crate::cli_util::CommandHelper;
 use crate::cli_util::RemoteBookmarkNamePattern;
 use crate::command_error::CommandError;
@@ -56,7 +56,7 @@ pub fn cmd_bookmark_untrack(
     let mut workspace_command = command.workspace_helper(ui)?;
     let repo = workspace_command.repo().clone();
     let mut symbols = Vec::new();
-    for (symbol, remote_ref) in find_remote_bookmarks(repo.view(), &args.names)? {
+    for (symbol, remote_ref) in find_trackable_remote_bookmarks(repo.view(), &args.names)? {
         if jj_lib::git::is_special_git_remote(symbol.remote) {
             // This restriction can be lifted if we want to support untracked @git
             // bookmarks.